requirements.txt
streamlit
openai
whisper
torchaudio

